+++ components/LoadingSpinner.tsx
@@ -1,4 +1,5 @@
 import React, { useState, useEffect } from 'react';
+import Image from 'next/image';
 import styles from '../styles/LoadingSpinner.module.css';
 
 const LoadingSpinner: React.FC = () => {
@@ -14,12 +15,20 @@ const LoadingSpinner: React.FC = () => {
 
   return (
     <div className={styles.spinnerContainer}>
-      <div className={styles.spinner}></div>
-      <p>Creating your magical story... This will take about 3-4 minutes, please wait.</p>
+      <div className={styles.spinner}>
+        <Image 
+          src="/SnugTales-Square-512.png" 
+          alt="SnugTales loading image"
+          width={400} // Increase size
+          height={400} // Increase size
+          layout="responsive"
+        />
+      </div>
+      <p className={styles.spinnerText}>Creating your magical story... This will take about 2-3 minutes, please wait.</p>
       <div className={styles.progressBar}>
         <div className={styles.progress} style={{ width: `${progress}%` }}></div>
       </div>
-      <p>{progress}% completed</p>
+      <p className={styles.progressText}>{progress}% completed</p>
     </div>
   );
 };